requirements.txt
asgiref==3.3.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
astroid==2.5.2; python_version >= "3.6" and python_version < "4.0"
certifi==2020.12.5; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cffi==1.14.5; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
chardet==4.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.6" and python_version < "4.0" and sys_platform == "win32" and python_full_version >= "3.5.0"
coreapi==2.3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
coreschema==0.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cryptography==3.4.7; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
defusedxml==0.7.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
django-cors-headers==3.7.0; python_version >= "3.6"
django-templated-mail==1.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
django==3.1.7; python_version >= "3.6"
djangorestframework-simplejwt==4.6.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
djoser==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
idna==2.10; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
isort==5.8.0; python_version >= "3.6" and python_version < "4.0"
itypes==1.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
jinja2==2.11.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
lazy-object-proxy==1.6.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.0"
markupsafe==1.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
mccabe==0.6.1; python_version >= "3.6" and python_version < "4.0"
oauthlib==3.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
psycopg2-binary==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pycparser==2.20; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pyjwt==2.0.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
pylint-django==2.4.2
pylint-plugin-utils==0.6
pylint==2.7.4; python_version >= "3.6" and python_version < "4.0"
python3-openid==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pytz==2021.1; python_version >= "3.6"
requests-oauthlib==1.3.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
requests==2.25.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
six==1.15.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-app-django==4.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-core==4.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.3.0"
uritemplate==3.0.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
urllib3==1.26.4; python_full_version >= "3.6.1" and python_version < "4" and python_full_version < "4.0.0"
wrapt==1.12.1; python_version >= "3.6" and python_version < "4.0"
